requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
arize-otel>=0.1.0 
openinference-instrumentation-openai